+++ src/lib/image_examiner.h
@@ -30,7 +30,7 @@ class ImageExaminer : public VideoExaminer
 public:
 	ImageExaminer (boost::shared_ptr<const Film>, boost::shared_ptr<const ImageContent>, boost::shared_ptr<Job>);
 
-	boost::optional<float> video_frame_rate () const;
+	boost::optional<double> video_frame_rate () const;
 	dcp::Size video_size () const;
 	Frame video_length () const {
 		return _video_length;
